updates react-router to v1.0.0-rc1
This presented a number of issues with configuring react-router. Most of them
were straight forward enough, but the one that I spent the most time fixing was
that we no longer use a `<RouteHandler/>` in our Main component to populate the
sub-views of the application. Instead we pass `{this.props.children}` and the
mysterious, impossible to debug, jerk of an error about `type.toUpperCase` goes
away.